warn about duplicate DAM metadata for same file on metadata migration
If multiple metadata for same files is encountered, we now print a
warning prompting the user to check migrated metadata of all listed
files (message includes file path and UIDs of sys_file and tx_dam
records).

This is for issue #62 again but still doesn't attempt to solve the
conflict, we leave it up to the user to check for any possible
errors.